requirements.txt
py-clob-client
python-dotenv
streamlit
requests
pandas
plotly
numpy
openai
pydantic
aiohttp
asyncio